[auth-frontend] Implement a basic login flow
This is missing some big chunks, like 2fa, and "skip 2fa for N days",
and non-ugly views, and an internal-api which actually handles these
endpoints.  But it's a start.

I know it's early days, but the interactor pattern looks good so far.
It was a bit of a nightmare when, at work, I tried to retrofit it onto
the govuk-account-manager-prototype; but using the pattern from the
get-go is quite nice.